quic: inline `Connection` into `muxer::Inner`
