+++ transports/quic/src/lib.rs
@@ -57,7 +57,6 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-mod connection;
 mod endpoint;
 mod error;
 mod muxer;
