+++ transports/quic/src/transport.rs
@@ -22,7 +22,6 @@
 //!
 //! Combines all the objects in the other modules to implement the trait.
 
-use crate::connection::Connection;
 use crate::endpoint::{Config, QuinnConfig, ToEndpoint};
 use crate::muxer::Inner;
 use crate::provider::Provider;
@@ -267,14 +266,7 @@ impl DialerState {
         async move {
             // Our oneshot getting dropped means the message didn't make it to the endpoint driver.
             let connection = tx.await.map_err(|_| Error::EndpointDriverCrashed)??;
-            let inner = Inner {
-                connection,
-                substreams: Default::default(),
-                poll_outbound_waker: None,
-                poll_inbound_waker: None,
-                poll_connection_waker: None,
-            };
-            let (peer, muxer) = Connecting::new(inner, timeout).await?;
+            let (peer, muxer) = Connecting::new(connection, timeout).await?;
 
             Ok((peer, muxer))
         }
@@ -307,7 +299,7 @@ struct Listener {
     listener_id: ListenerId,
 
     /// Channel where new connections are being sent.
-    new_connections_rx: mpsc::Receiver<Connection>,
+    new_connections_rx: mpsc::Receiver<Inner>,
     handshake_timeout: Duration,
 
     if_watcher: Option<IfWatcher>,
@@ -450,15 +442,8 @@ impl Stream for Listener {
                 Poll::Ready(Some(connection)) => {
                     let local_addr = socketaddr_to_multiaddr(connection.local_addr());
                     let send_back_addr = socketaddr_to_multiaddr(&connection.remote_addr());
-                    let inner = Inner {
-                        connection,
-                        substreams: Default::default(),
-                        poll_outbound_waker: None,
-                        poll_inbound_waker: None,
-                        poll_connection_waker: None,
-                    };
                     let event = TransportEvent::Incoming {
-                        upgrade: Connecting::new(inner, self.handshake_timeout),
+                        upgrade: Connecting::new(connection, self.handshake_timeout),
                         local_addr,
                         send_back_addr,
                         listener_id: self.listener_id,
